Perf improvements, part 2: don't reuse the CSSMinimiser object
Turns out it accumulates all the CSS it's seen so far within the
instance, so post N has to deal with N copies of the CSS.  Creating
a fresh instance each time makes this substantially faster.

We might be able to go faster by parsing the CSS once and then
only feeding in HTML, but that's a bit more fiddly.

On a small site, this takes build times from ~9.5s to ~1.5s.